requirements.txt
kafka-python
requests
psycopg2-binary
python-dotenv
pandas
sqlalchemy
logging

